# docs/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Volcanoes around the world v2",
        "description": "This API has been created to support assignment two in CAB230, Web Computing. It exposes a number of REST endpoints which implement CRUD operations on a database containing publicly available data about volcanoes collated by the Smithsonian Institution's Global Volcanism Program. The API endpoints and their usage are described in detail below.",
        "version": "1.0.0",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "apiteam@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "servers": [
        {
            "url": "http://localhost:3001"
        }
    ],
    "basePath": "/v2",
...
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "description": "Enter JWT Bearer token only"
            }
        },
...
        "responses": {
            "volcanoesError1": {
                "type": "object",
                "properties": {
                    "error": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Country is a required query parameter."
                    }
                }
            },
            "volcanoesError2": {
                "type": "object",
                "properties": {
                    "error": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Invalid populatedWithin query."
                    }
                }
            },
            "volcanoesError3": {
                "type": "object",
                "properties": {
                    "error": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "No query parameters."
                    }
                }
            },
            "volcanoError1": {
                "type": "object",
                "properties": {
                    "error": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Authorization header is malformed"
                    }
                }
            },
            "volcanoError2": {
                "type": "object",
                "properties": {
                    "error": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "JWT token has expired"
                    }
                }
            },
            "volcanoError3": {
                "type": "object",
                "properties": {
                    "error": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Invalid JWT token"
                    }
                }
            },
...
        }
    },
    "paths": {
        "/countries": {
            "get": {
                "tags": [
                    "Data"
                ],
                "summary": "",
                "description": "Returns a list of all countries that are associated with one or more volcanoes, ordered alphabetically.",
                "operationId": "countries",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "An array of countries. An example of the first five elements in the array is shown below.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    },
                                    "example": [
                                        "Algeria",
                                        "Antarctica",
                                        "Argentina",
                                        "Armenia",
                                        "Australia"
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid query parameters.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid query parameters. Query parameters are not permitted."
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/volcanoes": {
            "get": {
                "tags": [
                    "Data"
                ],
                "summary": "",
                "description": "Returns a list of volcanoes that are associated with the queried country. The country query parameter is required. The list can optionally be filtered by using the populatedWithin query parameter. This will return a filtered list of volcanoes that have at least one person living within the provided radius.",
                "operationId": "volcanoes",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "country",
                        "in": "query",
                        "description": "Name of country",
                        "required": true
                    },
                    {
                        "name": "populatedWithin",
                        "in": "query",
                        "description": "Distance within X km of a volcano",
                        "required": false,
                        "schema": {
                            "nullable": true,
                            "$ref": "#/components/populatedWithin"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "An array of objects containing id, name, country, region and subregion properties. An example of one object in the array is shown below.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Abu"
                                            },
                                            "country": {
                                                "type": "string",
                                                "example": "Japan"
                                            },
                                            "region": {
                                                "type": "string",
                                                "example": "Japan, Taiwan, Marianas"
                                            },
                                            "subregion": {
                                                "type": "string",
                                                "example": "Honshu"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Missing country query parameter or query parameters invalid. Click on 'Schema' below to see the possible error responses.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/responses/volcanoesError1"
                                        },
                                        {
                                            "$ref": "#/components/responses/volcanoesError2"
                                        },
                                        {
                                            "$ref": "#/components/responses/volcanoesError3"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
        "/volcano/{id}": {
            "get": {
                "tags": [
                    "Data"
                ],
                "summary": "",
                "description": "Returns an object containing data for the queried volcano. If a valid JWT token is sent in the header of the request, population data for 5km, 10km, 30km and 100km is also provided. To test this using Swagger, click the green 'Authorize' button at the top of this page to enter your JWT token. A JWT token can be obtained by logging in. The path parameter (id) is required.",
                "operationId": "volcano",
                "produces": [
                    "application/json"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the volcano",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns an object containing id, name, country, region, subregion, last eruption, summit, elevation, latitude and longitude data for the queried volcano. If a valid JWT token is sent in the header of the request, population data for 5km, 10km, 30km and 100km is also provided. An example of one object (with an authenticated request) is shown below.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Abu"
                                        },
                                        "country": {
                                            "type": "string",
                                            "example": "Japan"
                                        },
                                        "region": {
                                            "type": "string",
                                            "example": "Japan, Taiwan, Marianas"
                                        },
                                        "subregion": {
                                            "type": "string",
                                            "example": "Honshu"
                                        },
                                        "last_eruption": {
                                            "type": "string",
                                            "example": "6850 BCE"
                                        },
                                        "summit": {
                                            "type": "integer",
                                            "example": "641"
                                        },
                                        "elevation": {
                                            "type": "integer",
                                            "example": "2103"
                                        },
                                        "latitude": {
                                            "type": "string",
                                            "example": "34.5000"
                                        },
                                        "longitude": {
                                            "type": "string",
                                            "example": "131.6000"
                                        },
                                        "population_5km": {
                                            "type": "integer",
                                            "example": "3597"
                                        },
                                        "population_10km": {
                                            "type": "integer",
                                            "example": "9594"
                                        },
                                        "population_30km": {
                                            "type": "integer",
                                            "example": "117805"
                                        },
                                        "population_100km": {
                                            "type": "integer",
                                            "example": "4071152"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid parameters",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid query parameters. Query parameters are not permitted."
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized. Click on 'Schema' below to see the possible error responses.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/responses/volcanoError1"
                                        },
                                        {
                                            "$ref": "#/components/responses/volcanoError2"
                                        },
                                        {
                                            "$ref": "#/components/responses/volcanoError3"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Volcano ID not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Volcano with ID: 99999 not found."
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}